% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decennial.R
\name{census_datasets}
\alias{census_datasets}
\title{View Census Datasets}
\usage{
census_datasets(year)
}
\arguments{
\item{year}{Character or numeric, year of interest.}
}
\value{
Data frame with three columns.
}
\description{
View Census Datasets
}
