% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decennial.R
\name{get_dof}
\alias{get_dof}
\title{Extract P3 Data}
\usage{
get_dof(fips, dir)
}
\arguments{
\item{fips}{Character, length of three.}

\item{dir}{Full path, if NULL will use temporary directory.}
}
\value{
Data.frame with 6 columns.
}
\description{
Extract P3 Data
}
