% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decennial.R
\name{get_profile}
\alias{get_profile}
\title{Find County or State Code}
\usage{
get_profile(county, state = "CA")
}
\arguments{
\item{county}{Character, do not include word "county".}

\item{state}{Character, two letter abbreviation.}
}
\value{
Data.frame with four columns.
}
\description{
Find County or State Code
}
