% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners.R
\name{parse_label}
\alias{parse_label}
\title{Parse Variable Label}
\usage{
parse_label(x)
}
\arguments{
\item{x}{Expects "name" variable from variable list.}
}
\value{
Character.
}
\description{
Parse Variable Label
}
