% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decennial.R
\name{get_places}
\alias{get_places}
\title{Find CA Places Codes}
\usage{
get_places(county = "Orange", state = "CA")
}
\arguments{
\item{county}{Character, title casing.}

\item{state}{Character, length of two capitalized.}
}
\value{
Character, place codes.
}
\description{
Find CA Places Codes
}
