% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decennial.R
\name{vars_census}
\alias{vars_census}
\title{Census Variables}
\usage{
vars_census(year = 2020)
}
\arguments{
\item{year}{Character or integer, four characters in length.}
}
\value{
Data.frame, 3 columns.
}
\description{
Census Variables
}
